fix(rollup): remove options getter and make it a method
Apparently, there's a check that if a plugin has options, Rollup will call
it as an async function, and that was never the intention of this plugin
options dictionary, so I removed the getter and added a getOptions method.

This is not a breaking change because the plugin wasn't working up until now.

This is the line that was triggering the issue:
https://github.com/rollup/rollup/blob/master/src/rollup/rollup.ts#L126